test(core): move press tests to sibling file per command-module contract
